[#10964] refactor(idp-basic): Rename IdpRESTFeature and move REST extension test
Rename IdpRestFeature to IdpRESTFeature for naming consistency, relocate
TestIdpRestExtension to idp-basic, and drop server test dependency on the plugin.
